code cleanup and build fix
