+++ src/app/components/renderer/DocumentRenderer.tsx
