+++ src/app/components/renderer/SectionRenderer.tsx
