+++ src/app/components/utils/AnsiDisplay.tsx
