+++ src/app/components/views/ArticleMainPageRenderer.tsx
