+++ src/app/components/views/RootViewBlog.tsx
