--- src/app/pages/page.tsx
@@ -1,7 +1,0 @@
-import { filepathMatchPages } from "app/util/filepath";
-import { readNotebooksIndex } from "app/util/FsUtil";
-import { notFound } from "next/navigation";
-
-export default function BasePage (): never {
-  notFound();
-}
